Improve design
