+++ src/Angor.Avalonia.sln
@@ -14,6 +14,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AngorApp.Model", "AngorApp.Model\AngorApp.Model.csproj", "{5FEB5D7F-A9B3-47C7-B343-0126B90888E6}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Zafiro.Avalonia", "..\..\..\SuperJMN-Zafiro\Zafiro.Avalonia\src\Zafiro.Avalonia\Zafiro.Avalonia.csproj", "{33C93A6A-01ED-4C4B-BE2B-FC35BEBA1C73}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Zafiro.Avalonia.Dialogs", "..\..\..\SuperJMN-Zafiro\Zafiro.Avalonia\src\Zafiro.Avalonia.Dialogs\Zafiro.Avalonia.Dialogs.csproj", "{435A1549-15BC-4E6E-AF54-777FC57333F9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -32,6 +36,14 @@ Global
 		{5FEB5D7F-A9B3-47C7-B343-0126B90888E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5FEB5D7F-A9B3-47C7-B343-0126B90888E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5FEB5D7F-A9B3-47C7-B343-0126B90888E6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{33C93A6A-01ED-4C4B-BE2B-FC35BEBA1C73}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{33C93A6A-01ED-4C4B-BE2B-FC35BEBA1C73}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{33C93A6A-01ED-4C4B-BE2B-FC35BEBA1C73}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{33C93A6A-01ED-4C4B-BE2B-FC35BEBA1C73}.Release|Any CPU.Build.0 = Release|Any CPU
+		{435A1549-15BC-4E6E-AF54-777FC57333F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{435A1549-15BC-4E6E-AF54-777FC57333F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{435A1549-15BC-4E6E-AF54-777FC57333F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{435A1549-15BC-4E6E-AF54-777FC57333F9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
