+++ src/Angor/Avalonia/AngorApp/AddClassWhenUnmodifiedBehavior.cs
