+++ src/Angor/Avalonia/AngorApp/AngorApp.csproj
@@ -31,8 +31,6 @@
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="Zafiro.Avalonia"/>
-        <PackageReference Include="Zafiro.Avalonia.Dialogs"/>
 
     </ItemGroup>
 
@@ -52,6 +50,8 @@
     </ItemGroup>
 
     <ItemGroup>
+        <ProjectReference Include="..\..\..\..\..\..\SuperJMN-Zafiro\Zafiro.Avalonia\src\Zafiro.Avalonia.Dialogs\Zafiro.Avalonia.Dialogs.csproj" />
+        <ProjectReference Include="..\..\..\..\..\..\SuperJMN-Zafiro\Zafiro.Avalonia\src\Zafiro.Avalonia\Zafiro.Avalonia.csproj" />
         <ProjectReference Include="..\..\..\AngorApp.Model\AngorApp.Model.csproj"/>
     </ItemGroup>
 </Project>
