+++ src/Angor/Avalonia/AngorApp/Sections/Browse/Details/Invest/Amount/IAmountViewModel.cs
@@ -1,3 +1,4 @@
+using AngorApp.Model;
 using Zafiro.Avalonia.Controls.Wizards.Builder;
 
 namespace AngorApp.Sections.Browse.Details.Invest.Amount;
