+++ src/Angor/Avalonia/AngorApp/Sections/Browse/Details/ProjectDetailsViewModelDesign.cs
@@ -1,4 +1,5 @@
 using System.Windows.Input;
+using AngorApp.Model;
 
 namespace AngorApp.Sections.Browse.Details;
