+++ src/Angor/Avalonia/AngorApp/Sections/Browse/IProject.cs
@@ -1,20 +1,7 @@
+using AngorApp.Model;
+
 namespace AngorApp.Sections.Browse;
 
-public interface IProject
-{
-    public string Id { get; set; }
-    public string Name { get; }
-    public Uri Picture { get; }
-    public Uri Icon { get; }
-    public string ShortDescription { get; }
-    string BitcoinAddress { get; }
-    public decimal TargetAmount { get; }
-    public DateOnly StartingDate { get; }
-    IEnumerable<IStage> Stages { get; }
-    public string NpubKey { get; }
-    public string NpubKeyHex { get; }
-    public TimeSpan PenaltyDuration { get; }
-}
 
 public class ProjectDesign : IProject
 {
@@ -50,10 +37,3 @@ public class StageDesign : IStage
     public double Weight { get; set; }
 }
 
-public interface IStage
-{
-    DateOnly ReleaseDate { get; }
-    decimal Amount { get; }
-    int Index { get; }
-    double Weight { get; }
-}
