+++ src/AngorApp.Model/IBroadcastedTransaction.cs
