+++ src/AngorApp.Model/IStage.cs
@@ -1,0 +1,9 @@
+namespace AngorApp.Model;
+
+public interface IStage
+{
+    DateOnly ReleaseDate { get; }
+    decimal Amount { get; }
+    int Index { get; }
+    double Weight { get; }
+}
